requirements.txt
selenium
discord
cryptography.fernet
dotenv